+++ lib/capybara/cuprite/driver.rb
@@ -35,6 +35,8 @@ module Capybara
         @screen_size ||= DEFAULT_MAXIMIZE_SCREEN_SIZE
         @options[:save_path] ||= File.expand_path(Capybara.save_path) if Capybara.save_path
 
+        @options[:pending_connection_errors] = true unless @options.key?(:pending_connection_errors)
+
         # It's set for debug() to make devtools tab open correctly.
         @options[:browser_options] ||= {}
         unless @options[:browser_options][:"remote-allow-origins"]